Removed decimal factor from Tester

+++ MasterDataFlow.Trading/Data/Bar.cs
@@ -10,10 +10,10 @@ namespace MasterDataFlow.Trading.Data
     public class Bar
     {
         public DateTime Time { get; set; }
-        public double Open { get; set; }
-        public double Close { get; set; }
-        public double High { get; set; }
-        public double Low { get; set; }
-        public double Volume { get; set; }
+        public decimal Open { get; set; }
+        public decimal Close { get; set; }
+        public decimal High { get; set; }
+        public decimal Low { get; set; }
+        public decimal Volume { get; set; }
     }
 }
